requirements.txt
flask
flask-cors
transformers
torch
Flask
flask-cors
python-dotenv
groq

